lexiguard-backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
pydantic
google-cloud-dlp
google-cloud-translate==3.15.0
google-cloud-firestore
python-dotenv
PyPDF2
python-multipart
python-docx
reportlab
firebase-admin